+++ src/options/index.ts
@@ -3,4 +3,5 @@ export { template } from './template';
 export { license } from './license';
 export { org } from './org';
 export { remoteUrl } from './remote-url';
+export { templatize } from './templatize';
 export { verbose } from './verbose';