+++ src/utils/index.ts
@@ -5,3 +5,4 @@ export { isGitConfigured } from './git/is-git-configured';
 export { mergeTemplate } from './merge-template';
 export { replacePlaceholder } from './replace-placeholder';
 export { stage } from './git/stage';
+export { writeTemplateYaml } from './write-template-yaml';